Enstricten types more
We've got several different runtimes floating around - each folder has a tsconfig to ensure only the correct globals are available
