+++ package.json
@@ -20,9 +20,11 @@
   "homepage": "https://github.com/AsherFoster/WTFSIW#readme",
   "devDependencies": {
     "@asherfoster/eslint-config": "^3.0.0",
+    "@cloudflare/workers-types": "^3.4.0",
     "@types/node": "^16.11.26",
     "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.13",
+    "@types/service-worker-mock": "^2.0.1",
     "esbuild": "^0.14.25",
     "eslint": "^8.9.0",
     "eslint-config-prettier": "^8.5.0",
