+++ src/client/components/PreferenceAction.tsx
@@ -1,0 +1,7 @@
+import {RankingPreference} from '../../shared/clientapi/Scoring';
+
+const PreferenceAction = ({preference}: {preference: RankingPreference}) => {
+  return <a>Hi!</a>;
+};
+
+export default PreferenceAction;
