+++ src/scraper/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "lib": ["es2021"],
+    "types": ["node"]
+  }
+}
