+++ src/server/sentry.ts
@@ -1,4 +1,4 @@
 import * as Sentry from '@sentry/node';
-import {SENTRY_DSN} from '../config';
+import {SENTRY_DSN} from '../shared/config';
 
 Sentry.init({dsn: SENTRY_DSN});
