+++ src/shared/config.ts
@@ -1,5 +1,3 @@
-import * as path from 'path';
-
 export const ENVIRONMENT = process.env.NODE_ENV || 'development';
 const allowedEnvironments = ['development', 'production', 'test'];
 if (!allowedEnvironments.includes(ENVIRONMENT)) {
@@ -9,7 +7,6 @@ if (!allowedEnvironments.includes(ENVIRONMENT)) {
 }
 
 export const SENTRY_DSN = process.env.SENTRY_DSN;
-export const DATABASE_PATH = path.resolve(__dirname, '../data/db.sqlite');
 export const PORT = process.env.PORT || 8080;
 
 // Scoring algorithm parameters
